pyproject.toml
[project]
name = "bach-midi-generator-with-transformers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.16",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "attrs==25.3.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "docker-pycreds==0.4.0",
    "einops==0.8.1",
    "einx==0.3.0",
    "filelock==3.18.0",
    "frozendict==2.4.6",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "idna==3.10",
    "ipykernel>=6.29.5",
    "jinja2==3.1.6",
    "lightning-utilities==0.14.3",
    "loguru==0.7.3",
    "markupsafe==3.0.2",
    "mido==1.3.3",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "networkx==3.4.2",
    "numpy==2.2.4",
    "packaging==24.2",
    "platformdirs==4.3.7",
    "pretty-midi==0.2.10",
    "propcache==0.3.1",
    "protobuf==5.29.4",
    "psutil==7.0.0",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pytorch-lightning==2.5.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "sentry-sdk==2.26.1",
    "setproctitle==1.3.5",
    "six==1.17.0",
    "smmap==5.0.2",
    "sympy==1.13.1",
    "torch==2.6.0",
    "torchmetrics==1.7.1",
    "torchvision==0.21.0",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "wandb==0.19.9",
    "win32-setctime==1.2.0",
    "x-transformers==2.2.12",
    "yarl==1.20.0",
]
